// libs/@bambu/server-core/repositories/src/lib/notification-templates-repository/template-name.enum.ts
export enum TemplateNameEnum {
  VERIFY_LOGIN_OTP = 'VERIFY_LOGIN_OTP',
  LAUNCH_PLATFORM = 'LAUNCH_PLATFORM',
  SEND_INVESTMENT_PLAN_TO_LEAD = 'SEND_INVESTMENT_PLAN_TO_LEAD',
  RESET_PASSWORD_OTP = 'RESET_PASSWORD_OTP',
  SCHEDULE_APPOINTMENT = 'SCHEDULE_APPOINTMENT',
  PLEASE_AWAIT_APPOINTMENT = 'PLEASE_AWAIT_APPOINTMENT',
  INVESTOR_LOGIN_VERIFICATION_OTP = 'INVESTOR_LOGIN_VERIFICATION_OTP',
  WK_UPDATE_BANK_ACCOUNT_DETAILS = 'WK_UPDATE_BANK_ACCOUNT_DETAILS',
  WK_UPDATE_MODEL_PORTFOLIO_DETAILS = 'WK_UPDATE_MODEL_PORTFOLIO_DETAILS',
}